+++ app/src/main/res/layout/activity_add_wallet_item.xml
@@ -9,7 +9,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.ahmadrosid.dompetku.EditTransactionActivity">
+    tools:context="com.ahmadrosid.dompetku.transaction.EditTransactionActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
